+++ OpenDreamShared/Network/Messages/MsgNewAppearance.cs
@@ -11,18 +11,16 @@ namespace OpenDreamShared.Network.Messages;
 
 public sealed class MsgNewAppearance: NetMessage {
     public override MsgGroups MsgGroup => MsgGroups.EntityEvent;
-    public MsgNewAppearance() : this(MutableIconAppearance.Default) { }
-    public MsgNewAppearance(IBufferableAppearance appearance) => Appearance = appearance;
-    public IBufferableAppearance Appearance;
-    public int AppearanceId;
+
+    public MsgNewAppearance() : this(new ImmutableIconAppearance(MutableIconAppearance.Default, null)) {}
+    public MsgNewAppearance(ImmutableIconAppearance appearance) => Appearance = appearance;
+    public ImmutableIconAppearance Appearance;
 
     public override void ReadFromBuffer(NetIncomingMessage buffer, IRobustSerializer serializer) {
-        Appearance = new MutableIconAppearance();
-        AppearanceId = Appearance.ReadFromBuffer(buffer, serializer);
+        Appearance = new ImmutableIconAppearance(buffer, serializer);
     }
 
     public override void WriteToBuffer(NetOutgoingMessage buffer, IRobustSerializer serializer) {
-        AppearanceId = Appearance.GetHashCode();
         Appearance.WriteToBuffer(buffer,serializer);
     }
 }
